requirements.txt
pytz==2018.5
Django==2.1
djangorestframework==3.8.2
flake8==3.5.0
pip==18.0